requirements.txt
pandas
matplotlib
tkinterdnd2-pmgagne @ git+https://github.com/pmgagne/tkinterdnd2.git#egg=tkinterdnd2-pmgagne